pyproject.toml
[project]
name = "doc-parsing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "docling[vlm]>=2.72.0",
    "fsspec[http,s3]>=2025.10.0",
    "langdetect>=1.0.9",
    "pydantic>=2.12.5",
    "pypdf>=6.7.0",
    "pyyaml>=6.0.3",
    "rich>=14.3.2",
    "typer-slim>=0.21.1",
    "universal-pathlib>=0.3.9",
]

[project.scripts]
doc-parse = "doc_parsing.cli:app"

[project.entry-points."doc_parsing.adapters"]
docling = "doc_parsing.infrastructure.parsers.docling_adapter:adapter"
mock = "doc_parsing.infrastructure.parsers.mock_adapter:adapter"

[project.entry-points."doc_parsing.triage_policies"]
rules = "doc_parsing.infrastructure.triage.rules_policy:policy"

[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "prek>=0.3.3",
    "pytest>=9.0.2",
    "ruff>=0.15.1",
    "ty>=0.0.17",
]

[tool.uv]
package = true

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 88
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.ty.src]
include = ["src", "tests"]
